reorganize repo, convert to CMake project
Former build toolchain used automake and autoconf, both of which
are very unfriendly to cross-compilation and makes building nearly
impossible to accomplish for native Windows. It's also just a pain
in the ass to understand in general, especially when the target
audience are non-technical folks wishing to create their own tracks.

This commit reorganizes the repo into a defined structure, isolating
build artifacts from source files.